% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_names}
\alias{get_names}
\title{Get names from config \lifecycle{experimental}}
\usage{
get_names(..., type = valid__name_types(), as_character = FALSE)
}
\arguments{
\item{...}{[\link{character}] Name identifiers}

\item{type}{[\link{character}] Name type, on of \code{valid__name_types()}}

\item{as_character}{[\link{logical}] Return name as \code{name} (\code{FALSE}) or \code{character}
(\code{TRUE})}
}
\value{
[\link{list}] List of desired names (either as \code{names} or \code{character})
that can e.g. be used in \verb{dplyr::select(!!!<return_value>)} or the like
}
\description{
Get names from config, either as a list of \code{names} or \code{characters}.
Uses file path stored in environment variable \code{R_CONFIG_NAMES} (if set) or
\verb{./config.yml)} instead
}
