% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{map_names.character}
\alias{map_names.character}
\title{Map column names from/to internal/external value \lifecycle{experimental}}
\usage{
\method{map_names}{character}(cols, from = "external", to = "internal")
}
\arguments{
\item{cols}{[\link{list}] List containing the mapping information (\verb{list(internal = <value>, external = <value>)}) or list of such lists for multiple
columns. It is suggested that this is read from a YAML config file via
\code{confx::conf_get()}}

\item{from}{[\link{character}] List element from \code{cols} arg to describe the source
of the mapping}

\item{to}{[\link{character}] List element from \code{cols} arg to describe the target
of the mapping}
}
\value{
[\link{list}] List that can be used in conjunctin with the \verb{!!!} in
functions such as \code{dplyr::rename()} or \code{dplyr::select()}. The structure is
\verb{list(<to> = "<from>")}
}
\description{
Map column names from/to internal/external value \lifecycle{experimental}
}
