% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{map_names}
\alias{map_names}
\title{Map column names from/to internal/external value \lifecycle{experimental}}
\usage{
map_names(cols, from, to)
}
\arguments{
\item{cols}{Mapping information. The actual value depends on the method to be
used (\code{namap::map_names.list()} and
\code{namap::map_names.character()})
\code{confx::conf_get()}}

\item{from}{[\link{character}] List element from \code{cols} arg to describe the source
of the mapping}

\item{to}{[\link{character}] List element from \code{cols} arg to describe the target
of the mapping}
}
\value{
[\link{list}] List that can be used in conjunctin with the \verb{!!!} in
functions such as \code{dplyr::rename()} or \code{dplyr::select()}. The structure is
\verb{list(<to> = "<from>")}
}
\description{
Map column names from/to internal/external value \lifecycle{experimental}
}
