% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{trans__to_external_clean}
\alias{trans__to_external_clean}
\title{Clean up external name \lifecycle{experimental}}
\usage{
trans__to_external_clean(names)
}
\arguments{
\item{names}{[\link{list}] List with name specification (\verb{list(internal = <value>, external = <value>})}
}
\value{
[\link{character}] External name
}
\description{
Clean up external name \lifecycle{experimental}
}
